refactor(logging): move masking helper and extend guardrails
